ScreenPropTypes move to standalone index
