+++ Framework/Index/Root.ts
@@ -2,4 +2,3 @@
  * Exposed to end users, export core functions or comps here.
  */
 export { default as RootComp } from '../Root/RootComp';
-export { IScreenRouteProps, RootStackPropsList, IScreenProps } from '../Root/ScreenPropTypes';
